style(cpp26): drop more redundant casts; widen build_synthetic_apertures count
- sddp_state_io.cpp:196 — sim.phases().size() already returns size_t
- aperture_data_cache.cpp:91 — std::ssize(uid_str) over static_cast<ptrdiff_t>
- flow_right_lp.cpp:93 — OptInt::value_or already returns int
- simulation_lp.cpp:107 — strong-type PhaseIndex implicitly converts to int
- build_synthetic_apertures n_apertures: int → std::ptrdiff_t
  removes static_cast<int>(...size()) at three call sites in
  sddp_aperture_pass.cpp; replaced by std::ssize(...)

+++ source/flow_right_lp.cpp
@@ -90,7 +90,7 @@ Real resolve_block_fail_cost(Real sched_fail_cost,
 FlowRightLP::FlowRightLP(const FlowRight& pflow, const InputContext& ic)
     : ObjectLP<FlowRight>(pflow, ic, ClassName)
     , discharge(ic, ClassName, id(), std::move(flow_right().discharge))
-    , direction(static_cast<int>(flow_right().direction.value_or(-1)))
+    , direction(flow_right().direction.value_or(-1))
     , fail_cost_sched(ic, ClassName, id(), std::move(flow_right().fail_cost))
     , use_value_sched(ic, ClassName, id(), std::move(flow_right().use_value))
     , fmax_sched(ic, ClassName, id(), std::move(flow_right().fmax))
